scripts/requirements.txt
requests
pdfminer.six
python-docx
